+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file, as of versi
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.2] - 2021-08-26
+- add encoding.stringToBytes && encoding.bytesToString
+
+## [1.5.1] - 2021-08-26
+- add one param(gasUniPrice) in utils.tx.generateRawUserTransaction
+
 ## [1.5.0] - 2021-08-26
 - support contract.call_v2 method in rpc actions